회원탈퇴 + 게시글 작성 오류 수정

+++ dg/src/main/java/com/dajungdagam/dg/repository/WishListJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface WishListJpaRepository extends JpaRepository<Wishlist, Integer> {
+public interface WishListJpaRepository extends JpaRepository<Wishlist, Long> {
     Wishlist findByUser(User user);
 }
